Refactor: review carousel 컴포넌트 분리
